Move KerML validator tests to separate test file
- Created validator_test.rs for all validator tests
- Removed inline tests from validator.rs
- Follows project convention of keeping tests in separate files
