+++ crates/syster-base/src/semantic/adapters/kerml/mod.rs
@@ -2,3 +2,6 @@ mod helpers;
 mod population;
 pub mod validator;
 mod visitors;
+
+#[cfg(test)]
+mod validator_test;
